+++ packages/Webkul/Admin/src/DataGrids/Catalog/AttributeFamilyDataGrid.php
@@ -61,22 +61,26 @@ class AttributeFamilyDataGrid extends DataGrid
      */
     public function prepareActions()
     {
-        $this->addAction([
-            'icon'   => 'icon-edit',
-            'title'  => trans('admin::app.catalog.families.index.datagrid.catalog.families.index.datagrid.edit'),
-            'method' => 'GET',
-            'url'    => function ($row) {
-                return route('admin.catalog.families.edit', $row->id);
-            },
-        ]);
+        if (bouncer()->hasPermission('catalog.families.edit')) {
+            $this->addAction([
+                'icon'   => 'icon-edit',
+                'title'  => trans('admin::app.catalog.families.index.datagrid.catalog.families.index.datagrid.edit'),
+                'method' => 'GET',
+                'url'    => function ($row) {
+                    return route('admin.catalog.families.edit', $row->id);
+                },
+            ]);
+        }
 
-        $this->addAction([
-            'icon'   => 'icon-delete',
-            'title'  => trans('admin::app.catalog.families.index.datagrid.catalog.families.index.datagrid.delete'),
-            'method' => 'DELETE',
-            'url'    => function ($row) {
-                return route('admin.catalog.families.delete', $row->id);
-            },
-        ]);
+        if (bouncer()->hasPermission('catalog.families.edit')) {
+            $this->addAction([
+                'icon'   => 'icon-delete',
+                'title'  => trans('admin::app.catalog.families.index.datagrid.catalog.families.index.datagrid.delete'),
+                'method' => 'DELETE',
+                'url'    => function ($row) {
+                    return route('admin.catalog.families.delete', $row->id);
+                },
+            ]);
+        }
     }
 }
